pyproject.toml
[tool.poetry]
name = "pcoapi"
version = "0.0.2"
description = "Example Python project using best practices"
authors = ["John Hagen <johnthagen@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/johnthagen/python-blueprint"
documentation = "https://johnthagen.github.io/python-blueprint/"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
    # Include this classifier to prevent accidently publishing private code to PyPI.
    #   https://pypi.org/classifiers/
    "Private :: Do Not Upload",
]

[tool.poetry.dependencies]
# Some packages, such as scipy, constrain their upper bound of Python versions they support.
# Without also constraining the upper bound here, Poetry will not select those versions and will
# result in an old version being resolved/locked.
python = "^3.8, <3.12"
rich = "*"
typer = { version = "*", extras = ["all"] }
pypco = { git = "https://github.com/billdeitrick/pypco.git", tag = "v1.2.0"}
pydantic = "^2.0.1"

[tool.poetry.group.nox.dependencies]
nox-poetry = "*"

# TODO: Rename this to the "test" group when nox-poetry gains support for Poetry dependency
#   groups. https://github.com/cjolowicz/nox-poetry/issues/663
[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
python-dotenv = "*"

[tool.poetry.group.type_check.dependencies]
mypy = "*"
# As of mypy 0.900, mypy no longer bundles the stubs for third-party libraries that reside
# in the typeshed project. Add these "types-" packages here if you depend on them in
# requirements.in (e.g. types-requests).
# See: http://mypy-lang.blogspot.com/2021/06/mypy-0900-released.html
#      https://github.com/python/typeshed/tree/main/stubs

[tool.poetry.group.lint.dependencies]
ruff = "*"

[tool.poetry.group.fmt.dependencies]
black = "*"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "*"
mkdocs-htmlproofer-plugin = "*"
mkdocstrings = { version = "*", extras = ["python"] }
## Autodoc.
mkdocs-gen-files = "*"
mkdocs-literate-nav = "*"

#[tool.poetry.scripts]
#fact = "fact.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
strict = true
plugins = ["pydantic.mypy"]
# If certain strict config options are too pedantic for a project,
# disable them selectively here by setting to false.

[tool.ruff]
line-length = 99
target-version = "py38"
extend-select = [
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
extend-ignore = ["RUF005"]
src = ["src"]

[tool.ruff.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
# For non-src directory projects, explicitly set top level package names:
# known-first-party = ["my-app"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["typer.Argument"]

[tool.black]
line-length = 99
target-version = ["py38", "py39", "py310", "py311"]
# black will automatically exclude all files listed in .gitignore
# If you need to exclude additional folders, consider using extend-exclude to avoid disabling the
# default .gitignore behaviour.

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
]
markers = [
    "integration: marks as integration test",
    "notebooks: marks as notebook test",
    "gpu: marks as gpu test",
    "spark: marks tests which need Spark",
    "slow: marks tests as slow",
    "unit: fast offline tests",
]
xfail_strict = true
filterwarnings = [
    # When running tests, treat warnings as errors (e.g. -Werror).
    # See: https://docs.pytest.org/en/latest/reference/reference.html#confval-filterwarnings
    "error",
    # Add additional warning supressions as needed here. For example, if a third-party library
    # is throwing a deprecation warning that needs to be fixed upstream:
    # "ignore::DeprecationWarning:typer",
]

[tool.coverage.run]
branch = true